# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ZICK — A spiritual-wellness app backend providing auth, content, schedules, and subscriptions.",
        "title": "ZICK API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ZICK API Support",
            "email": "support@zick.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/motivations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to create a new motivation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motivation"
                ],
                "summary": "Create a motivation",
                "parameters": [
                    {
                        "description": "Motivation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.CreateMotivationReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.MotivationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/motivations/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to update an existing motivation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motivation"
                ],
                "summary": "Update a motivation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Motivation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Motivation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.UpdateMotivationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.MotivationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to delete an existing motivation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motivation"
                ],
                "summary": "Delete a motivation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Motivation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/motivations": {
            "get": {
                "description": "Retrieves a paginated list of motivation items. Supports search, sort, and pagination via query params.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motivation"
                ],
                "summary": "Get all motivations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by title or speaker name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort column; prefix with - for DESC (e.g. -created_at)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.PaginatedMotivationResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/motivations/{id}": {
            "get": {
                "description": "Retrieves details of a specific motivation along with related motivations.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motivation"
                ],
                "summary": "Get motivation details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Motivation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.MotivationDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gotickets_internal_domain_motivation_dto.CreateMotivationReq": {
            "type": "object",
            "required": [
                "description",
                "duration",
                "speaker_name",
                "thumbnail_url",
                "title",
                "video_url"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "string",
                    "maxLength": 20
                },
                "speaker_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "gotickets_internal_domain_motivation_dto.MotivationDetailsResponse": {
            "type": "object",
            "properties": {
                "motivation": {
                    "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.MotivationResponse"
                },
                "related": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.MotivationResponse"
                    }
                }
            }
        },
        "gotickets_internal_domain_motivation_dto.MotivationResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "speaker_name": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "gotickets_internal_domain_motivation_dto.PaginatedMotivationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gotickets_internal_domain_motivation_dto.MotivationResponse"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/gotickets_internal_querybuilder.Meta"
                }
            }
        },
        "gotickets_internal_domain_motivation_dto.UpdateMotivationReq": {
            "type": "object",
            "required": [
                "description",
                "duration",
                "speaker_name",
                "thumbnail_url",
                "title",
                "video_url"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "string",
                    "maxLength": 20
                },
                "speaker_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "gotickets_internal_httpresponse.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "gotickets_internal_querybuilder.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Limit is the number of items per page.",
                    "type": "integer"
                },
                "page": {
                    "description": "Page is the current page number (1-indexed).",
                    "type": "integer"
                },
                "total": {
                    "description": "Total is the total number of matching records.",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "TotalPages is the total number of pages.",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}